+++ lib/src/exports/reactive_mode.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:reactive_theme/src/private/theme_mutator.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 /// The `ReactiveMode` class provides a static method to check if the current theme mode is dark mode.
 class ReactiveMode {
@@ -12,4 +13,14 @@ class ReactiveMode {
   /// widget tree, such as accessing the current theme.
   static bool isDarkMode(BuildContext context) =>
       ThemeMutator.of(context).isDarkMode;
+
+  /// The function retrieves the saved theme mode from shared preferences and returns it as a ThemeMode
+  /// enum value.
+  ///
+  /// Returns:
+  ///   a `Future` object that resolves to a `ThemeMode` value or `null`.
+  static Future<ThemeMode?> getSavedThemeMode() async {
+    final prefs = await SharedPreferences.getInstance();
+    return (prefs.getBool('isdarkmode') ?? false) ? ThemeMode.dark : null;
+  }
 }