+++ lib/src/private/constants/constants.dart
@@ -1,5 +1,0 @@
-const String moonImg = 'assets/moon-fill.png';
-
-const String sunImg = 'assets/sun-fill.png';
-
-const String pkgname = 'reactive_theme';