+++ lib/src/private/theme_mutator.dart
@@ -1,16 +1,27 @@
 import 'package:flutter/material.dart';
-import 'package:reactive_theme/src/extension/reactive_theme_ext.dart';
+import 'package:reactive_theme/reactive_theme.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-/// The `StateChangerForThemeMutator` class is a StatefulWidget that manages the state for changing the
-/// theme mode of a `MaterialApp` widget.
+/// The `StateChangerForThemeMutator` class is a StatefulWidget that manages the state of a theme
+/// mutator widget and allows toggling between dark and light modes.
 class StateChangerForThemeMutator extends StatefulWidget {
-  const StateChangerForThemeMutator({super.key, required this.child});
+  const StateChangerForThemeMutator({
+    super.key,
+    required this.child,
+    required this.savedThemeMode,
+  });
 
-  /// The `final MaterialApp child;` is a property of the `StateChangerForThemeMutator` class. It
-  /// represents the child widget that will be wrapped by the `StateChangerForThemeMutator` widget. In
-  /// this case, it is the `MaterialApp` widget that will be modified to change the theme mode.
-  final MaterialApp child;
+  /// The `final ReactiveApp child;` is a parameter in the `StateChangerForThemeMutator` class
+  /// constructor. It represents the child widget that will be wrapped by the
+  /// `StateChangerForThemeMutator` widget. The `child` widget will receive the theme mode (either dark
+  /// or light) based on the state of the `isDarkMode` variable.
+  final ReactiveApp child;
+
+  /// The `final ThemeMode savedThemeMode;` is a parameter in the `StateChangerForThemeMutator` class
+  /// constructor. It represents the saved theme mode that will be used to initialize the `isDarkMode`
+  /// variable in the `initState` method. The `isDarkMode` variable determines whether the theme mode is
+  /// dark or light.
+  final ThemeMode savedThemeMode;
 
   @override
   State<StateChangerForThemeMutator> createState() =>
@@ -19,114 +30,59 @@ class StateChangerForThemeMutator extends StatefulWidget {
 
 class StateChangerForThemeMutatorState
     extends State<StateChangerForThemeMutator> {
-  /// The line `bool isDarkMode = false;` is declaring a boolean variable named `isDarkMode` and
-  /// initializing it with the value `false`. This variable is used to keep track of the current theme
-  /// mode (dark or light) in the `StateChangerForThemeMutator` widget.
-  bool isDarkMode = false;
+  /// The `late` keyword is used to declare a non-nullable variable that will be initialized later. In
+  /// this case, `late bool isDarkMode;` is declaring a boolean variable `isDarkMode` that will be
+  /// initialized in the `initState` method of the `StateChangerForThemeMutatorState` class. The
+  /// `isDarkMode` variable represents whether the theme mode is dark or light.
+  late bool isDarkMode;
 
-  /// The line `_key` is declaring a constant variable named `_key` and assigning
-  /// it the value `'isdarkmode'`. This variable is used as the key to store and retrieve the theme mode
-  /// (dark or light) in the shared preferences.
-  final _key = 'isdarkmode';
+  /// The function toggles the status of a variable called isDarkMode and saves the updated value in
+  /// SharedPreferences.
+  Future<void> toggleStatus() async {
+    final prefs = await SharedPreferences.getInstance();
+    setState(() => isDarkMode = !isDarkMode);
+    await prefs.setBool('isdarkmode', isDarkMode);
+  }
 
-  /// The initState function delays the execution of the readData function until after the current frame
-  /// is finished rendering.
+  /// The initState function sets the initial state of the isDarkMode variable based on the
+  /// savedThemeMode value.
   @override
   void initState() {
-    /// The line `Future.delayed(Duration.zero).then((_) => readData());` is delaying the execution of
-    /// the `readData()` function until after the current frame is finished rendering. This is done to
-    /// ensure that the `readData()` function is called after the widget has been fully initialized and
-    /// the context is available.
-    Future.delayed(Duration.zero).then((_) => readData());
+    /// The line `isDarkMode = ThemeMode.dark == widget.savedThemeMode` is comparing the value of
+    /// `widget.savedThemeMode` with `ThemeMode.dark`.
+    isDarkMode = ThemeMode.dark == widget.savedThemeMode;
     super.initState();
   }
 
-  /// The function reads data from SharedPreferences and updates the isDarkMode variable based on the
-  /// stored value or the platform's brightness.
-  Future<void> readData() async {
-    /// The line `prefs` is retrieving an instance of the
-    /// `SharedPreferences` class.
-    final prefs = await SharedPreferences.getInstance();
-
-    /// The `setState(() { ... })` method is used to update the state of the
-    /// `StateChangerForThemeMutator` widget. In this case, it is updating the `isDarkMode` variable
-    /// based on the stored value in the `SharedPreferences` or the platform's brightness.
-    setState(() {
-      isDarkMode = prefs.getBool(_key) ??
-          Theme.of(context).brightness == Brightness.dark;
-    });
-  }
-
-  /// The function toggles the status of a variable called isDarkMode and saves the updated value to
-  /// SharedPreferences.
-  Future<void> toggleStatus() async {
-    /// The line `prefs` is retrieving an instance of the
-    /// `SharedPreferences` class. `SharedPreferences` is a package that allows you to store and
-    /// retrieve key-value pairs persistently across app restarts. In this case, it is used to store and
-    /// retrieve the theme mode (dark or light) of the app. The `getInstance()` method returns a
-    /// `Future` that completes with the `SharedPreferences` instance, so `await` is used to wait for
-    /// the instance to be retrieved before continuing execution.
-    final prefs = await SharedPreferences.getInstance();
-
-    /// The line `setState(() => isDarkMode = !isDarkMode);` is updating the state of the
-    /// `StateChangerForThemeMutator` widget. It toggles the value of the `isDarkMode` variable by
-    /// negating its current value. This means that if `isDarkMode` is `true`, it will be set to
-    /// `false`, and if `isDarkMode` is `false`, it will be set to `true`. The `setState()` method is
-    /// used to notify the framework that the state has changed and triggers a rebuild of the widget
-    /// tree.
-    setState(() => isDarkMode = !isDarkMode);
-
-    /// `await prefs.setBool(_key, isDarkMode);` is saving the updated value of the `isDarkMode`
-    /// variable to the `SharedPreferences`.
-    await prefs.setBool(_key, isDarkMode);
-  }
-
-  /// The function builds a widget that allows for toggling between light and dark themes.
-  ///
-  /// Args:
-  ///   context (BuildContext): The BuildContext is a reference to the location of a widget within the
-  /// widget tree. It is used to access various properties and methods related to the current widget and
-  /// its ancestors.
-  ///
-  /// Returns:
-  ///   The `build` method is returning a `ThemeMutator` widget.
   @override
   Widget build(BuildContext context) {
+    /// The `return ThemeMutator(...)` statement is creating an instance of the `ThemeMutator` widget and
+    /// returning it as the result of the `build` method.
     return ThemeMutator(
+      /// The line `isDarkMode: isDarkMode,` is passing the value of the `isDarkMode` variable to the
+      /// `isDarkMode` parameter of the `ThemeMutator` widget. This allows the `ThemeMutator` widget to
+      /// access and use the value of `isDarkMode` within its own implementation.
       isDarkMode: isDarkMode,
 
-      /// In the `ThemeMutator` class, the `stateWidget: this` line is assigning the current instance of
-      /// the `StateChangerForThemeMutatorState` class to the `stateWidget` property of the
+      /// In the `ThemeMutator` class, the `stateWidget: this` parameter is assigning the current
+      /// instance of the `StateChangerForThemeMutatorState` class to the `stateWidget` parameter of the
       /// `ThemeMutator` widget.
       stateWidget: this,
 
-      /// The line `child: widget.child.themeChanger(themeMode: isDarkMode ? ThemeMode.light :
-      /// ThemeMode.dark),` is calling the `themeChanger` method on the `child` property of the
-      /// `StateChangerForThemeMutator` widget.
-      child: widget.child.themeChanger(
-        themeMode: isDarkMode ? ThemeMode.dark : ThemeMode.light,
-      ),
+      /// The line `child: widget.child.call(isDarkMode ? ThemeMode.dark : ThemeMode.light),` is calling
+      /// the `call` method on the `widget.child` object.
+      child: widget.child.call(isDarkMode ? ThemeMode.dark : ThemeMode.light),
     );
   }
 }
 
-/// The `ThemeMutator` class is an inherited widget that allows for changing the theme of the app and
-/// notifies its descendants when the theme changes.
 class ThemeMutator extends InheritedWidget {
-  /// The line `bool isDarkMode;` is declaring a boolean variable named `isDarkMode`. This
-  /// variable is used to keep track of the current theme mode (dark or light) in the
-  /// `StateChangerForThemeMutator` widget. It is marked as `final`, which means its value cannot be
-  /// changed once it is assigned.
+  /// The lines `final bool isDarkMode;` and `final StateChangerForThemeMutatorState stateWidget;` are
+  /// declaring two final variables in the `ThemeMutator` class.
   final bool isDarkMode;
 
-  /// The line `StateChangerForThemeMutatorState stateWidget;` is declaring a final variable named
-  /// `stateWidget` of type `StateChangerForThemeMutatorState`. This variable is used to store an
-  /// instance of the `StateChangerForThemeMutatorState` class, which represents the state of the
-  /// `StateChangerForThemeMutator` widget.
   final StateChangerForThemeMutatorState stateWidget;
 
-  /// The `const ThemeMutator({ ... })` is a constructor for the `ThemeMutator` class. It takes several
-  /// parameters:
   const ThemeMutator({
     super.key,
     required this.isDarkMode,